Terminate invasion cleanly when portspace is connected though other means
