+++ src/logic/map_objects/findimmovable.cc
@@ -19,6 +19,7 @@
 #include "logic/map_objects/findimmovable.h"
 
 #include "base/macros.h"
+#include "economy/economy.h"
 #include "economy/flag.h"
 #include "logic/map.h"
 #include "logic/map_objects/immovable.h"
@@ -74,25 +75,38 @@ bool FindImmovableAttackTarget::accept(const BaseImmovable& imm) const {
 // msites are visible.
 
 bool FindForeignMilitarysite::accept(const BaseImmovable& imm) const {
-	if (upcast(MilitarySite const, ms, &imm)) {
-		return &ms->owner() != &player;
-	}
-	return false;
-}
-
-bool FindImmovableByDescr::accept(const BaseImmovable& baseimm) const {
-	if (upcast(const Immovable, imm, &baseimm)) {
-		if (&imm->descr() == &descr) {
-			return true;
+	if (imm.descr().type() == MapObjectType::MILITARYSITE) {
+		if (upcast(MilitarySite const, ms, &imm)) {
+			return &ms->owner() != &player;
 		}
 	}
 	return false;
 }
 
+bool FindImmovableByDescr::accept(const BaseImmovable& baseimm) const {
+	if (&baseimm.descr() == &descr) {
+		return true;
+	}
+	return false;
+}
+
 bool FindFlagOf::accept(const BaseImmovable& baseimm) const {
-	if (upcast(const Flag, flag, &baseimm)) {
-		if (Building* building = flag->get_building()) {
+	if (baseimm.descr().type() == MapObjectType::FLAG) {
+		upcast(const Flag, flag, &baseimm);
+		if (Building* building = flag->get_building(); building != nullptr) {
 			if (finder.accept(*building)) {
+				return true;
+			}
+		}
+	}
+	return false;
+}
+
+bool FindFlagWithPlayersWarehouse::accept(const BaseImmovable& imm) const {
+	if (imm.descr().type() == MapObjectType::FLAG) {
+		upcast(const Flag, flag, &imm);
+		if (flag->get_owner() == &owner_) {
+			if (!flag->economy(wwWORKER).warehouses().empty()) {
 				return true;
 			}
 		}
